moved the map specific functions and structs
